Update to Product + DbInitializer

+++ laohaldusprojekt/laohaldusprojekt/Models/Product.cs
@@ -7,7 +7,7 @@ namespace laohaldusprojekt.Models
         public int Id { get; set; }
         public string TooteNimi { get; set; }
         public int Kogus { get; set; }
-        public decimal Hind { get; set; }
+        public double Hind { get; set; }
         public string Kategooria { get; set; }
     }
 }
